BugFix: Prevent weird prompting for writing unencrypted data to disk
After writing a SOPS encrypted secrets file, Neovim would follow it with
a prompt to write the unencrypted buffer data to the file instead, which
would overwrite the encrypted contents.

These changes correct that issue, and a few other minor tweaks.

Caveat note: We're intercepting the notify messages and only printing
them after the write is complete because otherwise Neovim prompts you to
press ENTER to continue after the write has completed.